+++ index.php
@@ -5,7 +5,7 @@ namespace sgkirby\Commentions;
 /**
  * Kirby 3 "Commentions" - Comments and Mentions Plugin
  *
- * @version   0.1.0
+ * @version   0.1.2
  * @author    Sebastian Greger <msg@sebastiangreger.net>
  * @copyright Sebastian Greger <msg@sebastiangreger.net>
  * @link      https://github.com/sebastiangreger/kirby3-sendmentions
@@ -20,21 +20,23 @@ require( __DIR__ . DS . 'helpers.php' );
 
 \Kirby::plugin('sgkirby/commentions', [
 
-    'api'          => require __DIR__ . '/config/api.php',
+    'api'     		=> require __DIR__ . '/config/api.php',
     
-    'blueprints' => [
+    'blueprints' 	=> [
         'fields/commentions' => __DIR__ . '/blueprints/fields/commentions.yml'
     ],
         
-    'sections' => [
+    'hooks' 		=> require __DIR__ . '/config/hooks.php',
+
+    'sections' 		=> [
         'commentions' => require __DIR__ . '/sections/commentions.php',
     ],
 
-    'routes'   => function ( $kirby ) {
+    'routes'   		=> function ( $kirby ) {
         return Commentions::endpointRoute( $kirby );
     },
     
-    'snippets' => [
+    'snippets' 		=> [
         'commentions-list' => __DIR__ . '/snippets/commentions-list.php',
         'commentions-form' => __DIR__ . '/snippets/commentions-form.php',
         'commentions-feedback' => __DIR__ . '/snippets/commentions-feedback.php',